+++ CHANGELOG.md
@@ -8,6 +8,10 @@
   - `preprocessHttpRequest(type, request) => Promise<HttpRequest>` to pre-process an HTTP request before the request is executed by the player
   - `preprocessHttpResponse(type, response) => Promise<HttpResponse>` to pre-process an HTTP response before it gets passed to the player
 
+### Changed
+
+- Update Bitmovin's native iOS SDK version to `3.64.0`
+
 ## [0.23.0] (2024-05-08)
 
 ### Changed